+++ docs/pyagentspec/source/changelog.rst
@@ -29,6 +29,13 @@ Improvements
   Note that this PR might break existing downstream LangGraph tracing span processors if they run in ab async environment;
   span processors will need to implement the async APIs (e.g. async def on_event_async) to properly use the async tracing mode.
 
+* **OCI GenAI model support in LangGraph adapter**
+
+  The LangGraph adapter now supports loading/exporting OCI GenAI models.
+  Users can now use models such ss Grok and Meta models available on the OCI GenAI service.
+  Install with `pip install pyagentspec[langgraph-full]` to access this feature.
+
+
 New features
 ^^^^^^^^^^^^